requirements.txt
django
gunicorn
django-heroku
django-cors-headers
pymongo
pymongo[srv]
dnspython
python-dotenv